Support classmap-authoritative for PHPStan Macro
