+++ src/Carbon/PHPStan/Macro.php
@@ -18,66 +18,10 @@ use ReflectionMethod;
 
 $method = new ReflectionMethod(BuiltinMethodReflection::class, 'getFileName');
 
-if ($method->hasReturnType()) {
-    final class Macro extends AbstractMacro
-    {
-        /**
-         * {@inheritdoc}
-         */
-        public function getFileName(): ?string
-        {
-            return $this->reflectionFunction->getFileName();
-        }
+require $method->hasReturnType()
+    ? __DIR__.'/../../../lazy/Carbon/PHPStan/MacroStrongType.php'
+    : __DIR__.'/../../../lazy/Carbon/PHPStan/MacroWeakType.php';
 
-        /**
-         * {@inheritdoc}
-         */
-        public function getStartLine(): ?int
-        {
-            return $this->reflectionFunction->getStartLine();
-        }
-
-        /**
-         * {@inheritdoc}
-         */
-        public function getEndLine(): ?int
-        {
-            return $this->reflectionFunction->getEndLine();
-        }
-    }
-
-    return;
-}
-
-final class Macro extends AbstractMacro
+final class Macro extends LazyMacro
 {
-    /**
-     * {@inheritdoc}
-     *
-     * @return string|false
-     */
-    public function getFileName()
-    {
-        return $this->reflectionFunction->getFileName();
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return int|false
-     */
-    public function getStartLine()
-    {
-        return $this->reflectionFunction->getStartLine();
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return int|false
-     */
-    public function getEndLine()
-    {
-        return $this->reflectionFunction->getEndLine();
-    }
 }
